CLASS-38(Blog-fastAPI)/Blog-fastAPI/requirements.txt
fastapi==0.115.6
uvicorn==0.32.1
SQLAlchemy==2.0.13
psycopg2==2.9.6
python-dotenv==1.0.0
inflect == 7.5.0
alembic==1.11.1
pydantic==1.10.0
passlib==1.7.4
email-validator==2.2.0
python-slugify==8.0.4
PyJWT==2.10.1
python-multipart==0.0.20